ACIBackEnd/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
psycopg2==2.9.10
PyJWT==2.9.0
sqlparse==0.5.1
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
